+++ packages/create-cloudflare/templates/remix/pages/templates/worker-configuration.d.ts
