+++ packages/create-cloudflare/templates/remix/pages/templates/wrangler.jsonc
