+++ packages/create-cloudflare/templates/remix/workers/templates/worker-configuration.d.ts
